+++ scripts/macros/2024/classFeatures/cleric/improvedBlessedStrikes.js
@@ -6,7 +6,7 @@ async function use({trigger: {entity: item}, workflow}) {
     if (workflowUtils.getCastLevel(workflow) != 0) return;
     let classIdentifier = itemUtils.getConfig(item, 'classIdentifier');
     if (workflow.item.system.sourceClass != classIdentifier) return;
-    if (activityUtils.isSpellActivity(workflow.activity)) return;
+    if (workflowUtils.isSustainedRoll(workflow)) return;
     let feature = activityUtils.getActivityByIdentifier(item, 'improvedPotentSpellcasting', {strict: true});
     if (!feature) return;
     let range = feature.range?.value ?? 60;